What's left are model-related utils, renamed and updated
